+++ controllers/contacts/removeContact.js
@@ -1,0 +1,18 @@
+const { Contact } = require("../models/contact");
+
+const { HttpError, ctrlWrapper } = require("../helpers");
+
+const removeContact = async (req, res) => {
+  const { contactId } = req.params;
+  const result = await Contact.findByIdAndRemove(contactId);
+
+  if (!result) {
+    throw HttpError(404, "Not found");
+  }
+
+  res.json({ message: "Contact deleted" });
+};
+
+module.exports = {
+  removeContact: ctrlWrapper(removeContact),
+};
